+++ src/test/kotlin/study/racing/RacingCarTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import study.racing.entity.RacingCar
+import study.racing.domain.entity.RacingCar
 
 class RacingCarTest {
