+++ src/test/kotlin/study/racing/RoundTest.kt
@@ -3,8 +3,8 @@ package study.racing
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import study.racing.entity.RacingCar
-import study.racing.entity.Round
+import study.racing.domain.entity.RacingCar
+import study.racing.domain.entity.Round
 
 class RoundTest {
